requirements.txt
django==2.1.1
django-extensions==2.1.3
django-filter==2.0.0
djangorestframework==3.8.2
pillow==5.3.0
requests==2.19.1
django-rest-auth
"django-rest-auth[with_social]"
djangorestframework-jwt
django-debug-toolbar
pillow
numpy
tensorflow
keras
haversine
drf-yasg
ipython==7.0.1